fix: keep non-Docker bg-process tests in standard E2E suite
