+++ playwright-manual.config.ts
@@ -26,7 +26,6 @@ export default defineConfig({
 			testDir: "./tests/e2e",
 			testMatch: [
 				"**/sandbox-recovery-docker*.spec.ts",
-				"**/bg-process-sandbox-guard*.spec.ts",
 			],
 		},
 	],
