+++ tests/e2e/bg-process-sandbox-guard.spec.ts
@@ -2,21 +2,13 @@
  * E2E tests for the BgProcessManager sandboxed-session guard.
  *
  * Verifies that sandboxed sessions without a containerId are refused
- * host-side execution (403), while sandboxed sessions with a containerId
- * and non-sandboxed sessions work normally.
+ * host-side execution (403) and non-sandboxed sessions work normally.
+ *
+ * The Docker-dependent test (sandboxed session WITH containerId) lives in
+ * sandbox-recovery-docker.spec.ts and runs via test:manual.
  */
-import { execFileSync } from "node:child_process";
 import { test, expect } from "./in-process-harness.js";
 import { readE2EToken, nonGitCwd } from "./e2e-setup.js";
-
-function isDockerAvailable(): boolean {
-	try {
-		execFileSync("docker", ["info"], { stdio: "ignore", timeout: 5000 });
-		return true;
-	} catch {
-		return false;
-	}
-}
 
 function adminFetch(baseURL: string, path: string, opts: RequestInit = {}) {
 	return fetch(`${baseURL}${path}`, {
@@ -53,33 +45,6 @@ test.describe("BgProcess Sandbox Guard", () => {
 		expect(bgRes.status).toBe(403);
 		const body = await bgRes.json();
 		expect(body.error).toBeTruthy();
-
-		// Cleanup
-		await adminFetch(gateway.baseURL, `/api/sessions/${id}`, { method: "DELETE" });
-	});
-
-	test("sandboxed session with containerId does not return 403", async ({ gateway }) => {
-		test.skip(!isDockerAvailable(), "Docker not available");
-		// Create a normal session
-		const res = await adminFetch(gateway.baseURL, "/api/sessions", {
-			method: "POST",
-			body: JSON.stringify({ cwd: nonGitCwd() }),
-		});
-		expect(res.status).toBe(201);
-		const { id } = await res.json();
-
-		// Manipulate session to be sandboxed WITH a containerId
-		const session = gateway.sessionManager.getSession(id);
-		session.sandboxed = true;
-		session.containerId = "fake-container-123";
-
-		// Attempt to create a bg-process — should NOT be 403
-		// (may fail due to fake container, but the sandbox guard should not block it)
-		const bgRes = await adminFetch(gateway.baseURL, `/api/sessions/${id}/bg-processes`, {
-			method: "POST",
-			body: JSON.stringify({ command: "echo test" }),
-		});
-		expect(bgRes.status).not.toBe(403);
 
 		// Cleanup
 		await adminFetch(gateway.baseURL, `/api/sessions/${id}`, { method: "DELETE" });
